pipe log data between the processes instea of tailing the file
All log data is now always sent to the parent process by the child.

The parent does all the writing to log or stdout. Now that we have
a stream of logs in the master process, it's easier to get that out
to the repl. Now live logs work all the time.