use abstract-blob-store/fs-blob-store
This sends all data that would be going directly to the filesystem for
mirring purposes through `options.blobstore` instead, which is
`fs-blob-store` by default.

Not all FS opertions have been replaced with blob-store operations. For
example, config files, index and error files are all accessed using
`fs`.

The `--blobstore` option is added, which provides a file to be required
that exports a blob-store.